Major fix
The UInt "bits" was being read as part of the "garc.btaf.nFiles" loop, but it did't belong there. the loop should have 3 uints, not 4. because of this, the reader became misaligned from the btaf entries. since writing this bad data would cause GARC Unpack to get stuck, it wroght a file for every otaf instead. This prevents getting stuck, but it dosn't write all the files. now that the btafs are fixed, GARC Unpack can extract all the files in the archive.